refactor: CafeService 관련 구조 변경
