--- src/main/java/com/example/demo/repository/cafe/CafeRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.example.demo.repository.cafe;
-
-public interface CafeRepositoryCustom {
-
-}
