--- src/main/java/com/example/demo/repository/cafe/CafeRepositoryCustomImpl.java
@@ -1,17 +1,0 @@
-package com.example.demo.repository.cafe;
-
-import javax.persistence.EntityManager;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-public class CafeRepositoryCustomImpl implements CafeRepositoryCustom {
-
-	private final EntityManager em;
-	private final JPAQueryFactory queryFactory;
-
-	public CafeRepositoryCustomImpl(EntityManager em) {
-		this.em = em;
-		this.queryFactory = new JPAQueryFactory(em);
-	}
-
-}
